GHC.Integer

Description

The Integer type.

This module exposes the portable Integer API. See GHC.Integer.GMP.Internals for the GMP-specific internal representation of Integer as well as optimized GMP-specific operations.

Documentation

data Integer

Arbitrary-precision integers.

Construct Integers

mkInteger

Arguments

:: Bool
sign of integer (True if non-negative)

-> [Int]
absolute value expressed in 31 bit chunks, least significant first

-> Integer

Construct Integer value from list of Ints.

This function is used by GHC for constructing Integer literals.

smallInteger :: Int# -> Integer

wordToInteger :: Word# -> Integer

Conversion to other integral types

integerToWord :: Integer -> Word#

integerToInt :: Integer -> Int#

Helpers for RealFloat type-class operations

encodeFloatInteger :: Integer -> Int# -> Float#

floatFromInteger :: Integer -> Float#

encodeDoubleInteger :: Integer -> Int# -> Double#

decodeDoubleInteger :: Double# -> (#Integer, Int##)

doubleFromInteger :: Integer -> Double#

Arithmetic operations

plusInteger :: Integer -> Integer -> Integer

minusInteger :: Integer -> Integer -> Integer

timesInteger :: Integer -> Integer -> Integer

negateInteger :: Integer -> Integer

absInteger :: Integer -> Integer

signumInteger :: Integer -> Integer

divModInteger :: Integer -> Integer -> (#Integer, Integer#)

divInteger :: Integer -> Integer -> Integer

modInteger :: Integer -> Integer -> Integer

quotRemInteger :: Integer -> Integer -> (#Integer, Integer#)

quotInteger :: Integer -> Integer -> Integer

remInteger :: Integer -> Integer -> Integer

Comparison predicates

eqInteger :: Integer -> Integer -> Bool

neqInteger :: Integer -> Integer -> Bool

leInteger :: Integer -> Integer -> Bool

gtInteger :: Integer -> Integer -> Bool

ltInteger :: Integer -> Integer -> Bool

geInteger :: Integer -> Integer -> Bool

compareInteger :: Integer -> Integer -> Ordering

eqInteger# :: Integer -> Integer -> Int#

Since: 0.5.1.0

neqInteger# :: Integer -> Integer -> Int#

Since: 0.5.1.0

leInteger# :: Integer -> Integer -> Int#

Since: 0.5.1.0

gtInteger# :: Integer -> Integer -> Int#

Since: 0.5.1.0

ltInteger# :: Integer -> Integer -> Int#

Since: 0.5.1.0

geInteger# :: Integer -> Integer -> Int#

Since: 0.5.1.0

Bit-operations

andInteger :: Integer -> Integer -> Integer

orInteger :: Integer -> Integer -> Integer

xorInteger :: Integer -> Integer -> Integer

complementInteger :: Integer -> Integer

shiftLInteger :: Integer -> Int# -> Integer

shiftRInteger :: Integer -> Int# -> Integer

testBitInteger :: Integer -> Int# -> Bool

Since: 0.5.1.0

Hashing

hashInteger :: Integer -> Int#

hashInteger returns the same value as fromIntegral, although in unboxed form. It might be a reasonable hash function for Integer, given a suitable distribution of Integer values.

Note: hashInteger is currently just an alias for integerToInt.